Xapian indexing: fix deadlocks, new MoinMoin.util.SubProcess module (see below)
Due to the usage of multiple pipes (stdout, stderr) to the subprocess (usually
a external filter program like pdftotext or antiword), deadlocks could occur
(especially when the filter program wrote lots of output to stderr, larger than
the buffer).

This was (hopefully) fixed as follows:

Python >= 2.4: use enhanced subclass of subprocess.Popen that implements a
timeout - if the process did not end when the timeout is reached, it will get
killed automatically.

Note: there is windows code to support the timeout functionality, but only the
linux code was tested. On windows with Python < 2.6, you need Python win32
extensions to support process termination.

Python 2.3: use popen2 to only get stdout (not stderr) - no support for:
 * timeout
 * stderr
 * rc
Please use Python >= 2.4 if you have issues with this.
We will drop the popen2 based code after requiring Python 2.4 with moin 1.9.