Fix SHA -> SSHA password hash upgrade for old user profiles (more details below).
If a user has a plaintext password that contains some non-ascii character and
his user profile was still an old profile (using a SHA hash), then moin failed
with an UnicodeError while trying to upgrade the user profile to SSHA.

As moin does this when the user logs in, the user could not log in.

Added some comment / docstring improvements to related code.